+++ app-react-birthday/src/components/ListTasks.jsx
@@ -1,7 +1,5 @@
 import { useEffect, useState } from "react";
-import { useDrag, useDrop } from "react-dnd";
-import toast from "react-hot-toast";
-import Task from "./Task";
+import Section from "./Section";
 
 const ListTasks = ({ tasks, setTasks }) => {
   const [todos, setTodos] = useState([]);
@@ -39,68 +37,3 @@ const ListTasks = ({ tasks, setTasks }) => {
 
 export default ListTasks;
 
-//ESTO ES UN COMPONENTE
-// las columnas de donde van a esta las listas
-const Section = ({ status, tasks, setTasks, todos, inProgress, closed }) => {
-  const [{ isOver }, drop] = useDrop(() => ({
-    accept: "task",
-    drop: (item) => addItemToSection(item.id),
-    collect: (monitor) => ({
-      isOver: !!monitor.isOver(),
-    }),
-  }));
-
-  let text = "Todo";
-  let bg = "bg-slate-500";
-  let tasksToMap = todos;
-
-  if (status === "inprogress") {
-    text = "In Progress";
-    bg = "bg-purple-500";
-    tasksToMap = inProgress;
-  }
-  if (status === "closed") {
-    text = "Closed";
-    bg = "bg-green-500";
-    tasksToMap = closed;
-  }
-
-  const addItemToSection = (id) => {
-    setTasks((prev) => {
-      const mTasks = prev.map((t) => {
-        if (t.id === id) {
-          return { ...t, status: status };
-        }
-        return t;
-      });
-      localStorage.setItem("tasks", JSON.stringify(mTasks));
-      toast("Task status change", { icon: "👌" });
-      return mTasks;
-      // console.log("prev",prev);
-    });
-    //console.log("dropped", id,status);
-  };
-  return (
-    <div
-      ref={drop}
-      className={`w-64 rounded-md p-2 ${isOver ? "bg-slate-400" : ""}`}
-    >
-      <Header text={text} bg={bg} count={tasksToMap.length} />
-      {tasksToMap.length > 0 &&
-        tasksToMap.map((task) => (
-          <Task key={task.id} task={task} tasks={tasks} setTasks={setTasks} />
-        ))}
-    </div>
-  );
-};
-///ESTO ES OTRO COMPONENTE
-const Header = ({ text, bg, count }) => {
-  return (
-    <div className={`${bg} flex items-center h-12 pl-4`}>
-      {text}
-      <div className="ml-2 bg-white w-5 h-5 text-black rounded-full flex  items-center justify-center ">
-        {count}
-      </div>
-    </div>
-  );
-};
